% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elbv2_operations.R
\name{delete_target_group}
\alias{delete_target_group}
\title{Deletes the specified target group}
\usage{
delete_target_group(
  TargetGroupArn = "string"
)
}
\arguments{
\item{TargetGroupArn}{[required] The Amazon Resource Name (ARN) of the target group.}
}
\description{
Deletes the specified target group.
}
\details{
You can delete a target group if it is not referenced by any actions. Deleting a target group also deletes any associated health checks.
}
\examples{
# This example deletes the specified target group.
delete_target_group(
  TargetGroupArn = "arn:aws:elasticloadbalancing:us-west-2:123456789012:targetgroup/my-targets/73e2d6bc24d8a067"
)

}
