% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elbv2_operations.R
\name{deregister_targets}
\alias{deregister_targets}
\title{Deregisters the specified targets from the specified target group}
\usage{
deregister_targets(TargetGroupArn, Targets)
}
\arguments{
\item{TargetGroupArn}{The Amazon Resource Name (ARN) of the target group.}

\item{Targets}{The targets. If you specified a port override when you registered a target, you must specify both the target ID and the port when you deregister it.}
}
\description{
Deregisters the specified targets from the specified target group. After the targets are deregistered, they no longer receive traffic from the load balancer.
}
\examples{
# This example deregisters the specified instance from the specified
# target group.
deregister_targets(
  TargetGroupArn = "arn:aws:elasticloadbalancing:us-west-2:123456789012:targetgroup/my-targets/73e2d6bc24d8a067",
  Targets = list(
    list(
      Id = "i-0f76fade"
    )
  )
)

}
