% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elbv2_operations.R
\name{describe_listeners}
\alias{describe_listeners}
\title{Describes the specified listeners or the listeners for the specified Application Load Balancer or Network Load Balancer}
\usage{
describe_listeners(LoadBalancerArn = NULL, ListenerArns = NULL,
  Marker = NULL, PageSize = NULL)
}
\arguments{
\item{LoadBalancerArn}{The Amazon Resource Name (ARN) of the load balancer.}

\item{ListenerArns}{The Amazon Resource Names (ARN) of the listeners.}

\item{Marker}{The marker for the next set of results. (You received this marker from a previous call.)}

\item{PageSize}{The maximum number of results to return with this call.}
}
\description{
Describes the specified listeners or the listeners for the specified Application Load Balancer or Network Load Balancer. You must specify either a load balancer or one or more listeners.
}
\examples{
# This example describes the specified listener.
describe_listeners(
  ListenerArns = list(
    "arn:aws:elasticloadbalancing:us-west-2:123456789012:listener/app/my-load-balancer/50dc6c495c0c9188/f2f7dc8efc522ab2"
  )
)

}
