% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elbv2_operations.R
\name{describe_rules}
\alias{describe_rules}
\title{Describes the specified rules or the rules for the specified listener}
\usage{
describe_rules(ListenerArn = NULL, RuleArns = NULL, Marker = NULL,
  PageSize = NULL)
}
\arguments{
\item{ListenerArn}{The Amazon Resource Name (ARN) of the listener.}

\item{RuleArns}{The Amazon Resource Names (ARN) of the rules.}

\item{Marker}{The marker for the next set of results. (You received this marker from a previous call.)}

\item{PageSize}{The maximum number of results to return with this call.}
}
\description{
Describes the specified rules or the rules for the specified listener. You must specify either a listener or one or more rules.
}
\section{Accepted Parameters}{
\preformatted{describe_rules(
  ListenerArn = "string",
  RuleArns = list(
    "string"
  ),
  Marker = "string",
  PageSize = 123
)
}
}

\examples{
# This example describes the specified rule.
\donttest{describe_rules(
  RuleArns = list(
    "arn:aws:elasticloadbalancing:us-west-2:123456789012:listener-rule/app/my-load-balancer/50dc6c495c0c9188/f2f7dc8efc522ab2/9683b2d02a6cabee"
  )
)}

}
