% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elbv2_operations.R
\name{modify_load_balancer_attributes}
\alias{modify_load_balancer_attributes}
\title{Modifies the specified attributes of the specified Application Load Balancer or Network Load Balancer}
\usage{
modify_load_balancer_attributes(LoadBalancerArn, Attributes)
}
\arguments{
\item{LoadBalancerArn}{The Amazon Resource Name (ARN) of the load balancer.}

\item{Attributes}{The load balancer attributes.}
}
\description{
Modifies the specified attributes of the specified Application Load Balancer or Network Load Balancer.
}
\details{
If any of the specified attributes can't be modified as requested, the call fails. Any existing attributes that you do not modify retain their current values.
}
\examples{
# This example enables deletion protection for the specified load
# balancer.
modify_load_balancer_attributes(
  Attributes = list(
    list(
      Key = "deletion_protection.enabled",
      Value = "true"
    )
  ),
  LoadBalancerArn = "arn:aws:elasticloadbalancing:us-west-2:123456789012:loadbalancer/app/my-load-balancer/50dc6c495c0c9188"
)

# This example changes the idle timeout value for the specified load
# balancer.
modify_load_balancer_attributes(
  Attributes = list(
    list(
      Key = "idle_timeout.timeout_seconds",
      Value = "30"
    )
  ),
  LoadBalancerArn = "arn:aws:elasticloadbalancing:us-west-2:123456789012:loadbalancer/app/my-load-balancer/50dc6c495c0c9188"
)

# This example enables access logs for the specified load balancer. Note
# that the S3 bucket must exist in the same region as the load balancer
# and must have a policy attached that grants access to the Elastic Load
# Balancing service.
modify_load_balancer_attributes(
  Attributes = list(
    list(
      Key = "access_logs.s3.enabled",
      Value = "true"
    ),
    list(
      Key = "access_logs.s3.bucket",
      Value = "my-loadbalancer-logs"
    ),
    list(
      Key = "access_logs.s3.prefix",
      Value = "myapp"
    )
  ),
  LoadBalancerArn = "arn:aws:elasticloadbalancing:us-west-2:123456789012:loadbalancer/app/my-load-balancer/50dc6c495c0c9188"
)

}
