% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elbv2_operations.R
\name{modify_rule}
\alias{modify_rule}
\title{Modifies the specified rule}
\usage{
modify_rule(RuleArn, Conditions = NULL, Actions = NULL)
}
\arguments{
\item{RuleArn}{The Amazon Resource Name (ARN) of the rule.}

\item{Conditions}{The conditions. Each condition specifies a field name and a single value.

If the field name is \code{host-header}, you can specify a single host name (for example, my.example.com). A host name is case insensitive, can be up to 128 characters in length, and can contain any of the following characters. You can include up to three wildcard characters.
\itemize{
\item A-Z, a-z, 0-9
\item - .
\item * (matches 0 or more characters)
\item ? (matches exactly 1 character)
}

If the field name is \code{path-pattern}, you can specify a single path pattern. A path pattern is case-sensitive, can be up to 128 characters in length, and can contain any of the following characters. You can include up to three wildcard characters.
\itemize{
\item A-Z, a-z, 0-9
\item \_ - . \$ / ~ " ' @ : +
\item & (using &)
\item * (matches 0 or more characters)
\item ? (matches exactly 1 character)
}}

\item{Actions}{The actions.

If the action type is \code{forward}, you can specify a single target group.

If the action type is \code{authenticate-oidc}, you can use an identity provider that is OpenID Connect (OIDC) compliant to authenticate users as they access your application.

If the action type is \code{authenticate-cognito}, you can use Amazon Cognito to authenticate users as they access your application.}
}
\description{
Modifies the specified rule.
}
\details{
Any existing properties that you do not modify retain their current values.

To modify the actions for the default rule, use ModifyListener.
}
\examples{
# This example modifies the condition for the specified rule.
modify_rule(
  Conditions = list(
    list(
      Field = "path-pattern",
      Values = list(
        "/images/*"
      )
    )
  ),
  RuleArn = "arn:aws:elasticloadbalancing:us-west-2:123456789012:listener-rule/app/my-load-balancer/50dc6c495c0c9188/f2f7dc8efc522ab2/9683b2d02a6cabee"
)

}
