% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elbv2_package.R
\docType{package}
\name{paws.elbv2-package}
\alias{paws.elbv2}
\alias{paws.elbv2-package}
\title{paws.elbv2: Elastic Load Balancing}
\description{
A load balancer distributes incoming traffic across
targets, such as your EC2 instances. This enables you to increase the
availability of your application. The load balancer also monitors the
health of its registered targets and ensures that it routes traffic
only to healthy targets. You configure your load balancer to accept
incoming traffic by specifying one or more listeners, which are
configured with a protocol and port number for connections from
clients to the load balancer. You configure a target group with a
protocol and port number for connections from the load balancer to the
targets, and with health check settings to be used when checking the
health status of the targets.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
