% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elbv2_operations.R
\name{remove_tags}
\alias{remove_tags}
\title{Removes the specified tags from the specified Elastic Load Balancing resource}
\usage{
remove_tags(
  ResourceArns = list(
    "string"
  ),
  TagKeys = list(
    "string"
  )
)
}
\arguments{
\item{ResourceArns}{[required] The Amazon Resource Name (ARN) of the resource.}

\item{TagKeys}{[required] The tag keys for the tags to remove.}
}
\description{
Removes the specified tags from the specified Elastic Load Balancing resource.
}
\details{
To list the current tags for your resources, use DescribeTags.
}
\examples{
# This example removes the specified tags from the specified load
# balancer.
remove_tags(
  ResourceArns = list(
    "arn:aws:elasticloadbalancing:us-west-2:123456789012:loadbalancer/app/my-load-balancer/50dc6c495c0c9188"
  ),
  TagKeys = list(
    "project",
    "department"
  )
)

}
