% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elbv2_operations.R
\name{set_ip_address_type}
\alias{set_ip_address_type}
\title{Sets the type of IP addresses used by the subnets of the specified Application Load Balancer or Network Load Balancer}
\usage{
set_ip_address_type(LoadBalancerArn, IpAddressType)
}
\arguments{
\item{LoadBalancerArn}{The Amazon Resource Name (ARN) of the load balancer.}

\item{IpAddressType}{The IP address type. The possible values are \code{ipv4} (for IPv4 addresses) and \code{dualstack} (for IPv4 and IPv6 addresses). Internal load balancers must use \code{ipv4}.}
}
\description{
Sets the type of IP addresses used by the subnets of the specified Application Load Balancer or Network Load Balancer.
}
\details{
Network Load Balancers must use \code{ipv4}.
}
\examples{

}
