% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elbv2_operations.R
\name{set_rule_priorities}
\alias{set_rule_priorities}
\title{Sets the priorities of the specified rules}
\usage{
set_rule_priorities(RulePriorities)
}
\arguments{
\item{RulePriorities}{The rule priorities.}
}
\description{
Sets the priorities of the specified rules.
}
\details{
You can reorder the rules as long as there are no priority conflicts in the new order. Any existing rules that you do not specify retain their current priority.
}
\examples{
# This example sets the priority of the specified rule.
set_rule_priorities(
  RulePriorities = list(
    list(
      Priority = 5L,
      RuleArn = "arn:aws:elasticloadbalancing:us-west-2:123456789012:listener-rule/app/my-load-balancer/50dc6c495c0c9188/f2f7dc8efc522ab2/1291d13826f405c3"
    )
  )
)

}
