% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elbv2_operations.R
\name{set_security_groups}
\alias{set_security_groups}
\title{Associates the specified security groups with the specified Application Load Balancer}
\usage{
set_security_groups(
  LoadBalancerArn = "string",
  SecurityGroups = list(
    "string"
  )
)
}
\arguments{
\item{LoadBalancerArn}{[required] The Amazon Resource Name (ARN) of the load balancer.}

\item{SecurityGroups}{[required] The IDs of the security groups.}
}
\description{
Associates the specified security groups with the specified Application Load Balancer. The specified security groups override the previously associated security groups.
}
\details{
You can't specify a security group for a Network Load Balancer.
}
\examples{
# This example associates the specified security group with the specified
# load balancer.
set_security_groups(
  LoadBalancerArn = "arn:aws:elasticloadbalancing:us-west-2:123456789012:loadbalancer/app/my-load-balancer/50dc6c495c0c9188",
  SecurityGroups = list(
    "sg-5943793c"
  )
)

}
