% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.elbv2_operations.R
\name{add_listener_certificates}
\alias{add_listener_certificates}
\title{Adds the specified certificate to the specified secure listener}
\usage{
add_listener_certificates(ListenerArn, Certificates)
}
\arguments{
\item{ListenerArn}{The Amazon Resource Name (ARN) of the listener.}

\item{Certificates}{The certificate to add. You can specify one certificate per call.}
}
\description{
Adds the specified certificate to the specified secure listener.
}
\details{
If the certificate was already added, the call is successful but the certificate is not added again.

To list the certificates for your listener, use DescribeListenerCertificates. To remove certificates from your listener, use RemoveListenerCertificates.
}
\examples{

}
